update range history API

+++ src/main/java/thc/domain/DailyStockQuote.java
@@ -2,4 +2,4 @@ package thc.domain;
 
 import java.time.LocalDate;
 
-public record DailyStockQuote(LocalDate date, double open, double close, double high, double low, int volume, double adjClose) {}
+public record DailyStockQuote(LocalDate date, double open, double close, double high, double low, long volume, double adjClose) {}
